installer: split pack/pdsc add/rm tests into separate files
Having all tests in a single file was starting to become cumbersome.
Splitting seems a fit now.
